--added BoardHandler
BoardHandler allows the multiplexing
of multiple boards together.

For instance, it allows to multiplex the player's board and the enemy's hidden board to allow for the player to see both his own board and the enemy's board that only shows hit and misses.

It contains the following functions:

--BoardHandler
This is the constructor, it allows you to set the parameters of the main board such
as its size, it's x and y offset and the index offset

--void create
This is the create function. It allows you to set the size of the boards that it will be taking in and the deafault marker that will show where there are no characters.

--void addBoard
This is the adding function. It allows you to add a single board into the main board. It allows you to select if there is indexing or not. Note that the board is mainly passed by value and that the single board does not affect the main board.

--add element
Although it should not be used often, it allows you to add a single character into the arbitrary space of the board. Note that it does not check for Out of index problems.

--getElement
This allows you to retrieve a single character from the main board. Note that it does not check for out of index problems.

--getCopyBoard
This returns a copy of the board. Note that this is pass by value.