Added CollisionHandler
This simple class handles detection of collisions inside the board with arbitrary characters and out of bounds detection.

Examples of when to use CollisionHandler:

When inputing a ship onto a board, call CollisionHandler on each spot the ship might occupy before altering the board. CollisionHandler will return true if it detects either a collision with another ship (that has the same character) and it reaches out of bounds of the board.

That's it!